requirements.txt
kagglehub
pandas
torch
torchvision
pillow
tqdm
accelerate>=0.16.0
transformers>=4.25.1
datasets
ftfy
Jinja2
diffusers>=0.14.0
git+https://github.com/huggingface/peft.git
wandb
# Image similarity search dependencies
faiss-cpu
streamlit
numpy